+++ src/main/java/test/game/model/entity/EntityTest.java
@@ -13,7 +13,7 @@ import test.game.model.world.WorldTestUtils;
 
 /**
  * Tests for Entities.
- *
+ * TODO Gabei shouldnt this be in defaultunittest?.
  * @author paladogabr
  */
 public class EntityTest {
@@ -47,7 +47,7 @@ public class EntityTest {
 
     World world = getWorld();
     player.setTargetUnit(enemy);
-    int prevHealth = enemy.getHealth();
+    double prevHealth = enemy.getHealth();
     for (int i = 0; i < 900; i++) {
       player.tick(GameModel.DELAY, world);
     }
